requirements.txt
asgiref==3.7.2
cloudinary==1.36.0
crispy-forms-materialize==0.2
dj-database-url==0.5.0
dj3-cloudinary-storage==0.0.6
Django==3.2
django-allauth==0.57.2
django-crispy-forms==2.0
django-summernote==0.8.20.0
gunicorn==21.2.0
oauthlib==3.2.2
psycopg2==2.9.9
PyJWT==2.8.0
python-dotenv==1.0.1
python3-openid==3.2.0
pytz==2024.1
requests-oauthlib==1.3.1
sqlparse==0.4.4
urllib3==1.26.18
whitenoise==5.3.0
